feat: save/read foods using Deno.KV
